requirements.txt
# Database connectivity
psycopg2-binary==2.9.9
sqlalchemy==2.0.36

# Data manipulation and analysis
pandas==2.2.3
numpy==2.1.3

# Machine Learning
scikit-learn==1.5.2
xgboost==1.7.6
imbalanced-learn==0.11.0
lifetimes==0.11.3

# Data Visualization
matplotlib==3.9.2
seaborn==0.13.2
plotly==6.1.2

# Dashboard
streamlit==1.45.1
streamlit-aggrid==0.3.4

# Utilities
python-dotenv==1.0.1
jupyter==1.0.0
ipykernel==6.25.0

# Logging and monitoring
loguru==0.7.3

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development tools
black==23.7.0
flake8==6.0.0
isort==5.12.0

# Additional dependencies
openpyxl==3.1.5
xlsxwriter==3.2.0
altair==5.5.0
pillow==11.0.0
tzdata==2024.2
pytz==2024.2 